@@ -9,78 +9,93 @@
     "start": "ts-node -r tsconfig-paths/register src/main.ts",
     "start:dev": "nodemon",
     "build": "rm -rf dist && tsc",
-    "prestart:prod": "rm -rf dist && tsc",
-    "start:prod": "node dist/main.js",
+    "prestart:prod": "npm run build",
+    "start:prod": "pm2 start ecosystem.config.js",
     "start:hmr": "node dist/server",
+    "repl": "rlwrap telnet localhost 5001",
     "test": "jest",
     "test:cov": "jest --coverage",
     "test:e2e": "jest --config ./test/jest-e2e.json",
-    "webpack": "webpack --config webpack.config.js"
+    "docker:build": "docker build -t smarterthings_api --file Dockerfile-arm --build-arg NODE_ENV=production .",
+    "docker:tag": "docker tag smarterthings_api registry.mattstrom.com/smarterthings_api",
+    "docker:push": "docker push registry.mattstrom.com/smarterthings_api",
+    "webpack": "webpack --config webpack.config.ts"
   },
   "dependencies": {
-    "@godaddy/terminus": "^2.4.1",
-    "@nestjs/common": "^5.2.1",
-    "@nestjs/core": "^5.2.1",
-    "@nestjs/microservices": "^5.2.1",
-    "@nestjs/passport": "^1.1.0",
-    "@nestjs/swagger": "^2.4.4",
-    "@nestjs/testing": "^5.2.1",
-    "@nestjs/typeorm": "^5.1.0",
-    "@nestjs/websockets": "^5.2.1",
+    "@godaddy/terminus": "^4.1.2",
+    "@nestjs/common": "^6.4.1",
+    "@nestjs/core": "^6.4.1",
+    "@nestjs/microservices": "^6.4.1",
+    "@nestjs/passport": "^6.1.0",
+    "@nestjs/platform-express": "^6.4.1",
+    "@nestjs/platform-socket.io": "^6.4.1",
+    "@nestjs/terminus": "^6.4.0",
+    "@nestjs/typeorm": "^6.1.2",
+    "@nestjs/websockets": "^6.4.1",
+    "app-root-path": "^2.2.1",
     "bcryptjs": "^2.4.3",
-    "class-transformer": "^0.1.9",
-    "class-validator": "^0.8.5",
-    "cookie-parser": "^1.4.3",
-    "core-js": "^2.5.7",
-    "dotenv": "^6.0.0",
-    "fastify-formbody": "^2.0.0",
-    "helmet": "^3.13.0",
-    "http-status-codes": "^1.3.0",
-    "mongodb": "^3.1.0",
-    "morgan": "^1.9.0",
-    "node-fetch": "^2.1.2",
+    "class-transformer": "^0.2.3",
+    "class-validator": "^0.9.1",
+    "config": "^3.1.0",
+    "convict": "^5.0.0",
+    "cookie-parser": "^1.4.4",
+    "fastify-formbody": "^3.1.0",
+    "helmet": "^3.18.0",
+    "http-signature": "^1.2.0",
+    "http-status-codes": "^1.3.2",
+    "morgan": "^1.9.1",
+    "node-fetch": "^2.6.0",
     "passport": "^0.4.0",
     "passport-jwt": "^4.0.0",
-    "reflect-metadata": "^0.1.12",
-    "request": "^2.87.0",
+    "reflect-metadata": "^0.1.13",
+    "request": "^2.88.0",
     "response-time": "^2.3.2",
-    "rxjs": "^6.0.0",
-    "simple-oauth2": "^2.2.0",
-    "typeorm": "^0.2.7",
-    "winston": "^3.0.0"
+    "rxjs": "^6.5.2",
+    "simple-oauth2": "^2.2.1",
+    "typeorm": "^0.2.18",
+    "typeorm-sqljs": "^1.2.2",
+    "uuid": "^3.3.2",
+    "winston": "^3.2.1"
   },
   "devDependencies": {
-    "@nestjs/cli": "^5.3.0",
-    "@types/bcryptjs": "^2.4.1",
+    "@nestjs/cli": "^6.5.0",
+    "@nestjs/testing": "^6.4.1",
+    "@types/app-root-path": "^1.2.4",
+    "@types/bcryptjs": "^2.4.2",
+    "@types/config": "^0.0.34",
+    "@types/convict": "^4.2.1",
     "@types/cookie-parser": "^1.4.1",
-    "@types/core-js": "^2.5.0",
-    "@types/dotenv": "^4.0.3",
-    "@types/express": "^4.16.0",
-    "@types/helmet": "0.0.38",
-    "@types/jest": "^21.1.8",
+    "@types/express": "^4.17.0",
+    "@types/helmet": "^0.0.43",
+    "@types/jest": "^24.0.15",
     "@types/morgan": "^1.7.35",
-    "@types/node": "^10.7.0",
-    "@types/node-fetch": "^2.1.2",
-    "@types/passport": "^0.4.5",
+    "@types/node": "^12.0.12",
+    "@types/node-fetch": "^2.3.7",
+    "@types/passport": "^1.0.0",
     "@types/passport-jwt": "^3.0.1",
-    "@types/request": "^2.47.1",
-    "@types/response-time": "^2.3.2",
-    "@types/simple-oauth2": "^1.1.1",
-    "@types/supertest": "^2.0.4",
-    "@types/winston": "^2.3.9",
-    "jest": "^21.2.1",
-    "nodemon": "^1.14.1",
-    "prettier": "^1.11.1",
-    "supertest": "^3.0.0",
-    "ts-jest": "^21.2.4",
-    "ts-loader": "^4.4.2",
-    "ts-node": "^7.0.1",
-    "tsconfig-paths": "^3.5.0",
-    "tslint": "5.3.2",
-    "typescript": "^3.0.1",
-    "webpack": "^4.2.0",
-    "webpack-cli": "^2.0.13",
-    "webpack-node-externals": "^1.6.0"
+    "@types/request": "^2.48.1",
+    "@types/response-time": "^2.3.3",
+    "@types/simple-oauth2": "^2.2.1",
+    "@types/socket.io": "^2.1.2",
+    "@types/supertest": "^2.0.8",
+    "@types/uuid": "^3.4.5",
+    "jest": "^24.8.0",
+    "nodemon": "^1.19.1",
+    "pm2": "^3.5.1",
+    "prettier": "^1.18.2",
+    "supertest": "^4.0.2",
+    "ts-jest": "^24.0.2",
+    "ts-loader": "^6.0.4",
+    "ts-node": "^8.3.0",
+    "tsconfig-paths": "^3.8.0",
+    "tslint": "^5.18.0",
+    "typescript": "^3.5.2",
+    "webpack": "^4.35.2",
+    "webpack-cli": "^3.3.5",
+    "webpack-node-externals": "^1.7.2"
+  },
+  "_moduleAliases": {
+    "~": "./dist"
   },
   "jest": {
     "moduleFileExtensions": [
