@@ -3,57 +3,71 @@
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve",
+    "start": "ng serve --host 0.0.0.0 --disableHostCheck true",
+    "start:server": "browser-sync --config support/bs-config.js start",
     "build": "ng build",
+    "build:support": "tsc --build",
     "test": "ng test",
     "lint": "ng lint",
-    "e2e": "ng e2e"
+    "e2e": "ng e2e",
+    "docker:build": "docker build -t smarterthings_ui --file Dockerfile-arm .",
+    "docker:tag": "docker tag smarterthings_ui registry.mattstrom.com/smarterthings_ui",
+    "docker:push": "docker push registry.mattstrom.com/smarterthings_ui"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.0.4",
-    "@angular/cdk": "^7.0.4",
-    "@angular/common": "^7.0.4",
-    "@angular/compiler": "^7.0.4",
-    "@angular/core": "^7.0.4",
-    "@angular/forms": "^7.0.4",
-    "@angular/http": "^7.0.4",
-    "@angular/material": "^7.0.4",
-    "@angular/platform-browser": "^7.0.4",
-    "@angular/platform-browser-dynamic": "^7.0.4",
-    "@angular/platform-server": "^7.0.4",
-    "@angular/pwa": "^0.10.6",
-    "@angular/router": "^7.0.4",
-    "@angular/service-worker": "^7.0.4",
-    "@auth0/angular-jwt": "^2.0.0",
-    "@ngxs/store": "^3.2.0",
-    "core-js": "^2.5.4",
+    "@angular/animations": "^7.1.3",
+    "@angular/cdk": "^7.1.1",
+    "@angular/common": "^7.1.3",
+    "@angular/compiler": "^7.1.3",
+    "@angular/core": "^7.1.3",
+    "@angular/forms": "^7.1.3",
+    "@angular/http": "^7.1.3",
+    "@angular/material": "^7.1.1",
+    "@angular/platform-browser": "^7.1.3",
+    "@angular/platform-browser-dynamic": "^7.1.3",
+    "@angular/platform-server": "^7.1.3",
+    "@angular/pwa": "^0.11.3",
+    "@angular/router": "^7.1.3",
+    "@angular/service-worker": "^7.1.3",
+    "@auth0/angular-jwt": "^2.1.0",
+    "@fortawesome/angular-fontawesome": "^0.3.0",
+    "@fortawesome/fontawesome-svg-core": "^1.2.10",
+    "@fortawesome/free-solid-svg-icons": "^5.6.1",
+    "@ngxs/storage-plugin": "^3.3.2",
+    "@ngxs/store": "^3.3.2",
+    "core-js": "^2.6.0",
+    "hammerjs": "^2.0.8",
     "http-status-codes": "^1.3.0",
     "js-cookie": "^2.2.0",
+    "moment": "^2.23.0",
     "rxjs": "^6.3.3",
-    "socket.io-client": "^2.1.1",
+    "socket.io-client": "^2.2.0",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.6.8",
-    "@angular/cli": "^7.0.6",
-    "@angular/compiler-cli": "^7.0.4",
-    "@angular/language-service": "^7.0.4",
-    "@types/jasmine": "~2.8.6",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/js-cookie": "^2.1.0",
-    "@types/node": "~8.9.4",
+    "@angular-devkit/build-angular": "~0.11.3",
+    "@angular/cli": "^7.1.3",
+    "@angular/compiler-cli": "^7.1.3",
+    "@angular/language-service": "^7.1.3",
+    "@ngxs/devtools-plugin": "^3.3.2",
+    "@types/browser-sync": "0.0.42",
+    "@types/jasmine": "~3.3.1",
+    "@types/jasminewd2": "~2.0.6",
+    "@types/js-cookie": "^2.2.0",
+    "@types/node": "~10.12.14",
     "@types/socket.io-client": "^1.4.32",
+    "browser-sync": "^2.26.3",
     "codelyzer": "~4.5.0",
     "http-server": "^0.11.1",
-    "jasmine-core": "~2.99.1",
+    "jasmine-core": "~3.3.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^3.0.0",
+    "karma": "^3.1.3",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.0",
-    "karma-jasmine": "~1.1.1",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "^5.4.0",
+    "karma-coverage-istanbul-reporter": "~2.0.4",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
+    "protractor": "^5.4.1",
     "ts-node": "~7.0.1",
     "tslint": "~5.11.0",
     "typescript": "~3.1.6"
