@@ -1,13 +1,11 @@
 import { Module } from '@nestjs/common';
-import { APP_INTERCEPTOR } from '@nestjs/core';
-import { TypeOrmModule } from '@nestjs/typeorm';
 
 import { AppController } from './app.controller';
 import { SecuritySystemController, TriggerController } from './controllers';
+import { DatabaseModule } from './modules';
 import { AuthModule } from './modules/auth';
-import { LoggingModule, MorganInterceptor } from './modules/logging';
+import { LoggingModule } from './modules/logging';
 import { EventsGateway } from './services';
-import { ServerId, RedirectUrl, SmartThingsAppUrl } from './types';
 
 
 @Module({
@@ -18,30 +16,11 @@ import { ServerId, RedirectUrl, SmartThingsAppUrl } from './types';
 	],
 	imports: [
 		AuthModule,
-		LoggingModule.forRoot(),
-		TypeOrmModule.forRoot({
-			type: 'mongodb',
-			host: process.env.TYPEORM_HOST || 'localhost',
-			port: 27017,
-			database: 'smarterthings',
-			entities: [
-				__dirname + '/entities/**/*.entity{.ts,.js}',
-			],
-			synchronize: true
-		} as any)
+		DatabaseModule,
+		LoggingModule.forRoot()
 	],
 	providers: [
-		EventsGateway,
-		{ provide: RedirectUrl, useValue: process.env.REDIRECT_URL },
-		{ provide: ServerId, useValue: process.env.SERVER_ID },
-		{
-			provide: SmartThingsAppUrl,
-			useValue: process.env.SMARTTHINGS_TOKEN_HOST
-		},
-		{
-			provide: APP_INTERCEPTOR,
-			useClass: MorganInterceptor,
-		}
+		EventsGateway
 	]
 })
 export class AppModule { }
