@@ -3,9 +3,9 @@ declare namespace NodeJS {
 		API_URL: string;
 		ENDPOINT_URL: string;
 		JWT_SECRET: string;
+		PORT: number;
 		REDIRECT_URL: string;
 		SERVER_ID: string;
-		TYPEORM_HOST: string;
 
 		SMARTTHINGS_CLIENT_ID: string;
 		SMARTTHINGS_CLIENT_SECRET: string;
