@@ -1,33 +1,25 @@
-require('dotenv').config(); // tslint:disable-line:no-var-requires
-
-import './polyfills';
-import * as terminus from '@godaddy/terminus';
-import { ValidationPipe } from '@nestjs/common';
+import { createTerminus } from '@godaddy/terminus';
+import { Logger, ValidationPipe } from '@nestjs/common';
 import { NestFactory } from '@nestjs/core';
-import { SwaggerModule, DocumentBuilder } from '@nestjs/swagger';
-import * as cookieParser from 'cookie-parser';
-import * as helmet from 'helmet';
-import * as responseTime from 'response-time';
+import { NestExpressApplication } from '@nestjs/platform-express';
+import cookieParser from 'cookie-parser';
+import helmet from 'helmet';
+import responseTime from 'response-time';
 
+import config from './configuration';
 import { AppModule } from './app.module';
 import { logger } from './modules/logging/winston';
+import { ReplModule } from './services/repl.module';
+import { ReplServer } from './services/repl-server.service';
 
 
+declare const module: any;
+
+const port: number = (config.port) || 4567;
 let ready: boolean = false;
 
 async function bootstrap() {
-	const app = await NestFactory.create(AppModule);
-
-	const options = new DocumentBuilder()
-		.setTitle('SmarterThings')
-		.setDescription('SmarterThings API description')
-		.setVersion('1.0')
-		.addTag('smartthings')
-		.build();
-
-	const document = SwaggerModule.createDocument(app, options);
-
-	SwaggerModule.setup('docs', app, document);
+	const app = await NestFactory.create<NestExpressApplication>(AppModule);
 
 	app.use(cookieParser());
 	app.use(helmet());
@@ -40,11 +32,20 @@ async function bootstrap() {
 		allowedHeaders: ['Accept', 'Authorization', 'Content-Type']
 	});
 
-	await app.listen(4567, () => {
+	await app.listen(port, () => {
 		ready = true;
 	});
 
-	terminus(app.getHttpServer(), {
+	const repl = await NestFactory.createMicroservice(ReplModule, {
+		strategy: new ReplServer({
+			context: app
+		}),
+		logger: new Logger()
+	});
+
+	repl.listen(() => {});
+
+	createTerminus(app.getHttpServer(), {
 		signal: 'SIGINT',
 		healthChecks: {
 			'/health/liveness': async () => Promise.resolve(),
@@ -59,6 +60,11 @@ async function bootstrap() {
 			logger.info('System shutdown.');
 		}
 	});
+
+	if (module.hot) {
+		module.hot.accept();
+		module.hot.dispose(() => app.close());
+	}
 }
 
 bootstrap();
