@@ -1,8 +1,8 @@
-import { Inject, Injectable, UnauthorizedException } from '@nestjs/common';
+import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { PassportStrategy } from '@nestjs/passport';
 import { ExtractJwt, Strategy } from 'passport-jwt';
 
-import { SecretKey } from '../../types';
+import config from '../../configuration';
 import { AuthService } from './auth.service';
 
 
@@ -14,15 +14,14 @@ export interface JwtPayload {
 export class JwtStrategy extends PassportStrategy(Strategy) {
 
 	constructor(
-		private readonly authService: AuthService,
-		@Inject(SecretKey) private readonly secretKey: string
+		private readonly authService: AuthService
 	) {
 		super({
 			jwtFromRequest: ExtractJwt.fromExtractors([
 				ExtractJwt.fromUrlQueryParameter('token'),
 				ExtractJwt.fromAuthHeaderAsBearerToken()
 			]),
-			secretOrKey: secretKey
+			secretOrKey: config.jwtSecret
 		});
 	}
 
