@@ -1,2 +1,3 @@
-export const Logger = Symbol('Logger');
-export const MorganProvider = Symbol('MorganProvider');
+export const Logger: unique symbol = Symbol('Logger');
+export const MorganProvider: unique symbol = Symbol('MorganProvider');
+export const NestLogger: unique symbol = Symbol('NestLogger');
