@@ -1,6 +1,1 @@
-export const EndpointUrl = Symbol('EndpointUrl');
-export const OauthModuleOptions = Symbol('OauthModuleOptions');
-export const RedirectUrl = Symbol('RedirectUrl');
-export const SecretKey = Symbol('SecretKey');
-export const ServerId = Symbol('ServerId');
-export const SmartThingsAppUrl = Symbol('SmartThingsAppUrl');
+export const OauthModuleOptions: unique symbol = Symbol('OauthModuleOptions');
