@@ -1,26 +1,31 @@
+import { HTTP_INTERCEPTORS } from '@angular/common/http';
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { RouterModule } from '@angular/router';
 import { ServiceWorkerModule } from '@angular/service-worker';
 import { JwtModule } from '@auth0/angular-jwt';
+import { NgxsReduxDevtoolsPluginModule } from '@ngxs/devtools-plugin';
+import { NgxsStoragePluginModule } from '@ngxs/storage-plugin';
 import { NgxsModule } from '@ngxs/store';
 
 import { environment } from '../environments/environment';
 import { PublicSideComponent } from './components/public-side/public-side.component';
 import { AppComponent } from './app.component';
+import { SecuredSideComponent } from './components/secured-side/secured-side.component';
 import {
 	AuthGuard,
 	AuthModule,
 	ConnectComponent,
 	LoginComponent,
 	RegisterComponent,
-	SmartThingsAuthGuard
+	SmartThingsAuthGuard, TokenInterceptor
 } from './modules/auth';
-import { WebSocketService } from './modules/keypad/services';
+import { SensorsService, WebSocketService } from './services';
 import { UiComponentsModule } from './modules/ui-components';
 import { IdentityService } from './services';
-import { LoadingStatus } from './state';
+import { LoadingStatusInterceptor } from './services/loading-status.interceptor';
+import { Sensors, KeycodeState, LoadingStatus, SecuritySystem, DeviceState } from './state';
 import { tokenGetter } from './token-getter';
 import { ApiUrlToken, WsUrlToken } from './tokens';
 
@@ -28,7 +33,8 @@ import { ApiUrlToken, WsUrlToken } from './tokens';
 @NgModule({
 	declarations: [
 		AppComponent,
-		PublicSideComponent
+		PublicSideComponent,
+		SecuredSideComponent
 	],
 	imports: [
 		BrowserModule.withServerTransition({appId: 'serverApp'}),
@@ -43,17 +49,22 @@ import { ApiUrlToken, WsUrlToken } from './tokens';
 			}
 		}),
 		NgxsModule.forRoot([
-			LoadingStatus
-		]),
+			DeviceState,
+			LoadingStatus,
+			SecuritySystem,
+			KeycodeState,
+			Sensors
+		], { developmentMode: !environment.production }),
+		NgxsReduxDevtoolsPluginModule.forRoot(),
+		// NgxsStoragePluginModule.forRoot(),
 		RouterModule.forRoot([
 			{
 				path: '',
-				redirectTo: '/keypad',
+				redirectTo: '/home',
 				pathMatch: 'full'
 			},
 			{
 				path: '',
-				component: PublicSideComponent,
 				children: [
 					{
 						path: 'connect',
@@ -71,18 +82,31 @@ import { ApiUrlToken, WsUrlToken } from './tokens';
 				]
 			},
 			{
-				path: 'keypad',
-				loadChildren: './modules/keypad/keypad.module#KeypadModule',
+				path: 'home',
+				loadChildren: './modules/home/home.module#HomeModule',
 				canLoad: [AuthGuard],
 				canActivate: [SmartThingsAuthGuard]
-			}
+			},
+			{
+				path: 'settings',
+				loadChildren: './modules/settings/settings.module#SettingsModule',
+				canLoad: [AuthGuard],
+				canActivate: [SmartThingsAuthGuard]
+			},
+
 		])
 	],
 	providers: [
 		IdentityService,
+		SensorsService,
 		WebSocketService,
-		{ provide: ApiUrlToken, useValue: '/api' },
-		{ provide: WsUrlToken, useValue: `ws://${location.host}` }
+		{ provide: ApiUrlToken, useValue: environment.apiUrl },
+		{ provide: WsUrlToken, useValue: environment.wsUrl },
+		{
+			provide: HTTP_INTERCEPTORS,
+			useClass: LoadingStatusInterceptor,
+			multi: true
+		}
 	],
 	bootstrap: [AppComponent]
 })
