@@ -1,5 +1,7 @@
 import { Component, OnInit } from '@angular/core';
+import { Select } from '@ngxs/store';
 import { Observable } from 'rxjs';
+import { KeycodeState } from '../../../../state';
 
 import { KeycodeService, SecuritySystemService } from '../../services';
 
@@ -12,24 +14,20 @@ export type CodeStatus = 'valid' | 'invalid' | 'unknown';
 	styleUrls: ['./keycode-display.component.scss']
 })
 export class KeycodeDisplayComponent implements OnInit {
-	code$: Observable<any>;
-	codeStatus: CodeStatus = 'unknown';
 
-	constructor(
-		private keycodeService: KeycodeService,
-		private securitySystem: SecuritySystemService
-	) {
-		this.code$ = this.keycodeService.code$;
-	}
+	@Select(KeycodeState.getCode)
+	code$: Observable<string>;
+
+	constructor() {}
 
 	ngOnInit() {
-		this.securitySystem.validCode$
-			.subscribe((value) => {
-				this.codeStatus = (value !== null)
-					? (value === true)
-						? 'valid'
-						: 'invalid'
-					: 'unknown';
-			});
+		// this.securitySystem.validCode$
+		// 	.subscribe((value) => {
+		// 		this.codeStatus = (value !== null)
+		// 			? (value === true)
+		// 				? 'valid'
+		// 				: 'invalid'
+		// 			: 'unknown';
+		// 	});
 	}
 }
