@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 
 @Component({
-	selector: 'keypad-container',
+	selector: 'smt-keypad-container',
 	templateUrl: './keypad-container.component.html',
 	styleUrls: ['./keypad-container.component.scss']
 })
