@@ -1,26 +1,32 @@
-import { AfterViewInit, Component, OnInit, OnDestroy, QueryList, ViewChildren } from '@angular/core';
+import { DOCUMENT } from '@angular/common';
+import { AfterViewInit, Component, OnInit, OnDestroy, QueryList, ViewChildren, Inject } from '@angular/core';
+import { Store } from '@ngxs/store';
 import { fromEvent, Subscription } from 'rxjs';
 import { filter, map } from 'rxjs/operators';
 
+import { PressKey } from '../../../../state';
 import { Blop } from '../../../../utils/sound';
-import { KeycodeService } from '../../services';
 import { KeypadButtonComponent } from '../keypad-button/keypad-button.component';
 
 
 @Component({
-	selector: 'numpad',
+	selector: 'smt-numpad',
 	templateUrl: './keypad.component.html',
 	styleUrls: ['./keypad.component.scss']
 })
 export class NumpadComponent implements OnInit, AfterViewInit, OnDestroy {
 	@ViewChildren(KeypadButtonComponent) buttons: QueryList<KeypadButtonComponent>;
 
+	private subscription: Subscription;
 	private subscriptions: Subscription[];
 
-	constructor(private keycodeService: KeycodeService) { }
+	constructor(
+		private readonly store: Store,
+		@Inject(DOCUMENT) private readonly document: Document
+	) { }
 
 	ngOnInit() {
-		fromEvent(document, 'keypress')
+		this.subscription = fromEvent(this.document, 'keypress')
 			.pipe(
 				filter((event: KeyboardEvent) => {
 					return event.keyCode >= 48 && event.keyCode <= 57;
@@ -31,7 +37,7 @@ export class NumpadComponent implements OnInit, AfterViewInit, OnDestroy {
 				const key = value.toString();
 
 				this.playSound(key);
-				this.keycodeService.input$.next(key);
+				this.store.dispatch(new PressKey(key));
 			});
 	}
 
@@ -44,6 +50,7 @@ export class NumpadComponent implements OnInit, AfterViewInit, OnDestroy {
 	}
 
 	ngOnDestroy() {
+		this.subscription.unsubscribe();
 		this.subscriptions.forEach((subscription) => {
 			subscription.unsubscribe();
 		});
@@ -51,7 +58,7 @@ export class NumpadComponent implements OnInit, AfterViewInit, OnDestroy {
 
 	onClick(key: string) {
 		this.playSound(key);
-		this.keycodeService.input$.next(key);
+		this.store.dispatch(new PressKey(key));
 	}
 
 	private playSound(value: string) {
