@@ -1,15 +1,14 @@
 import { Component, OnInit } from '@angular/core';
 
 @Component({
-  selector: 'smt-toolbar-progress-bar',
-  templateUrl: './toolbar-progress-bar.component.html',
-  styleUrls: ['./toolbar-progress-bar.component.scss']
+	selector: 'smt-toolbar-progress-bar',
+	templateUrl: './toolbar-progress-bar.component.html',
+	styleUrls: ['./toolbar-progress-bar.component.scss']
 })
 export class ToolbarProgressBarComponent implements OnInit {
 
-  constructor() { }
+	constructor() {}
 
-  ngOnInit() {
-  }
+	ngOnInit() {}
 
 }
